requirements.txt
Authlib==1.2.1
fastapi==0.103.1
openai==0.27.8
pydantic==1.10.10
python-dotenv==1.0.0
Requests==2.31.0
uvicorn==0.23.2
validators==0.20.0